# Improve performance of card rendering parser